+++ LudenWebAPI/Application/Abstractions/Interfaces/Services/IBillService.cs
@@ -5,7 +5,7 @@ namespace Application.Abstractions.Interfaces.Services
 {
     public interface IBillService : IGenericService<Bill>
     {
-        Task<Bill> CreateBillAsync(int userId, decimal totalAmount, BillStatus status);
-        Task<IEnumerable<Bill>> GetBillsByUserIdAsync(int userId);
+        Task<Bill> CreateBillAsync(ulong userId, decimal totalAmount, BillStatus status);
+        Task<IEnumerable<Bill>> GetBillsByUserIdAsync(ulong userId);
     }
 }
