+++ LudenWebAPI/Application/Abstractions/Interfaces/Services/IFileService.cs
@@ -5,14 +5,14 @@ namespace Application.Abstractions.Interfaces.Services
 {
     public interface IFileService : IGenericService<FileEntity>
     {
-        Task<PhotoFile> UploadUserAvatarAsync(int userId, Stream fileStream, string fileName, string contentType, long fileSize);
-        Task<ProductFile> UploadProductFileAsync(int productId, Stream fileStream, string fileName, string contentType, long fileSize, string fileType);
-        Task<PhotoFile?> GetPhotoFileByIdAsync(int id);
-        Task<ProductFile?> GetProductFileByIdAsync(int id);
-        Task<IEnumerable<ProductFile>> GetProductFilesAsync(int productId);
-        Task<PhotoFile?> GetUserAvatarAsync(int userId);
-        Task DeletePhotoFileAsync(int id);
-        Task DeleteProductFileAsync(int id);
+        Task<PhotoFile> UploadUserAvatarAsync(ulong userId, Stream fileStream, string fileName, string contentType, long fileSize);
+        Task<ProductFile> UploadProductFileAsync(ulong productId, Stream fileStream, string fileName, string contentType, long fileSize, string fileType);
+        Task<PhotoFile?> GetPhotoFileByIdAsync(ulong id);
+        Task<ProductFile?> GetProductFileByIdAsync(ulong id);
+        Task<IEnumerable<ProductFile>> GetProductFilesAsync(ulong productId);
+        Task<PhotoFile?> GetUserAvatarAsync(ulong userId);
+        Task DeletePhotoFileAsync(ulong id);
+        Task DeleteProductFileAsync(ulong id);
         string GetFileUrl(string path);
     }
 }
